Fix Trust gauge

+++ Assets/Scripts/Modifiers/Modifier.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 [System.Serializable]
 public class Modifier
 {
-    public int AddedValue = 0;
-    public int OneShotValue = 0;
+    public float AddedValue = 0;
+    public float OneShotValue = 0;
 }
